refactor e2e tests, move random_refs and api_client out
