+++ tests/e2e/test_api.py
@@ -1,16 +1,6 @@
 import pytest
-import requests
-
-from allocation import config
-from ..random_refs import random_sku, random_batchref, random_orderid
-
-
-def post_to_add_batch(ref, sku, qty, eta):
-    url = config.get_api_url()
-    r = requests.post(
-        f"{url}/add_batch", json={"ref": ref, "sku": sku, "qty": qty, "eta": eta}
-    )
-    assert r.status_code == 201
+from ..random_refs import random_batchref, random_orderid, random_sku
+from . import api_client
 
 
 @pytest.mark.usefixtures("postgres_db")
@@ -20,24 +10,27 @@ def test_happy_path_returns_201_and_allocated_batch():
     earlybatch = random_batchref(1)
     laterbatch = random_batchref(2)
     otherbatch = random_batchref(3)
-    post_to_add_batch(laterbatch, sku, 100, "2011-01-02")
-    post_to_add_batch(earlybatch, sku, 100, "2011-01-01")
-    post_to_add_batch(otherbatch, othersku, 100, None)
-    data = {"orderid": random_orderid(), "sku": sku, "qty": 3}
+    api_client.post_to_add_batch(laterbatch, sku, 100, "2011-01-02")
+    api_client.post_to_add_batch(earlybatch, sku, 100, "2011-01-01")
+    api_client.post_to_add_batch(otherbatch, othersku, 100, None)
 
-    url = config.get_api_url()
-    r = requests.post(f"{url}/allocate", json=data)
+    response = api_client.post_to_allocate(random_orderid(), sku, qty=3)
 
-    assert r.status_code == 201
-    assert r.json()["batchref"] == earlybatch
+    assert response.status_code == 201
+    assert response.json()["batchref"] == earlybatch
 
 
 @pytest.mark.usefixtures("postgres_db")
 @pytest.mark.usefixtures("restart_api")
 def test_unhappy_path_returns_400_and_error_message():
     unknown_sku, orderid = random_sku(), random_orderid()
-    data = {"orderid": orderid, "sku": unknown_sku, "qty": 20}
-    url = config.get_api_url()
-    r = requests.post(f"{url}/allocate", json=data)
-    assert r.status_code == 400
-    assert r.json()["message"] == f"Invalid sku {unknown_sku}"
+
+    response = api_client.post_to_allocate(
+        orderid,
+        unknown_sku,
+        qty=20,
+        expect_success=False,
+    )
+
+    assert response.status_code == 400
+    assert response.json()["message"] == f"Invalid sku {unknown_sku}"
